Adds a CuImage payload backed by a the memory pool in core.
